requirements.txt
Keras-Preprocessing==1.1.2
Flask==2.2.2
Flask-Cors==3.0.10
imageio==2.22.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.8.0
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
seaborn==0.12.1
termcolor==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1


